split menus into Local and Remote sourced-menus
