+++ views/menus/menus/base.js
@@ -1,135 +1,48 @@
 // @flow
 import React from 'react'
 import {View, Navigator, Platform} from 'react-native'
-import LoadingView from '../../components/loading'
 import FancyMenu from '../parts/fancy-menu'
 import {FilterToolbar} from '../filter/toolbar'
 
 import type {TopLevelViewPropsType} from '../../types'
-import {TopLevelViewPropTypes} from '../../types'
 import type momentT from 'moment'
-import moment from 'moment-timezone'
-const CENTRAL_TZ = 'America/Winnipeg'
-import findIndex from 'lodash/findIndex'
 import uniq from 'lodash/uniq'
 import values from 'lodash/values'
 import map from 'lodash/map'
 
 import type {
-  BonAppMenuInfoType,
   StationMenuType,
   MenuItemContainerType,
-  DayPartMenuType,
-  DayPartsCollectionType,
-  BonAppCafeInfoType,
   MasterCorIconMapType,
 } from '../types'
 
 import type {FilterSpecType} from '../filter/types'
 
-
 export class BaseMenuView extends React.Component {
-  static propTypes = {
-    cafeId: React.PropTypes.string.isRequired,
-    cafeInfo: React.PropTypes.object.isRequired,
-    cafeMenu: React.PropTypes.object.isRequired,
-    now: React.PropTypes.instanceOf(moment).isRequired,
-    ...TopLevelViewPropTypes,
-  }
-
   state: {|
-    menus: StationMenuType[],
-    foodItems: MenuItemContainerType,
-    message: string|null,
-    isFiltered: bool,
     filters: FilterSpecType[],
-    mealName: string,
   |} = {
-    menus: [],
-    foodItems: {},
-    message: null,
-    isFiltered: false,
-    mealName: '',
     filters: [],
   }
 
   componentDidMount() {
-    this.parseData()
+    this.buildFilters()
   }
 
   props: TopLevelViewPropsType & {
     now: momentT,
-    cafeId?: string,
-    cafeInfo?: BonAppCafeInfoType,
-    cafeMenu?: BonAppMenuInfoType,
-    menuStations?: StationMenuType[],
-    menuItems?: MenuItemContainerType,
-    menuLabel?: string,
-    menuCorIcons?: MasterCorIconMapType,
-  }
-
-  findMenu(dayparts: DayPartsCollectionType, now: momentT): void|DayPartMenuType {
-    if (!dayparts.length || !dayparts[0].length) {
-      return
-    }
-
-    if (dayparts[0].length === 1) {
-      return dayparts[0][0]
-    }
-
-    const times = dayparts[0].map(({starttime, endtime}) => ({
-      startTime: moment.tz(starttime, 'H:mm', true, CENTRAL_TZ),
-      endTime: moment.tz(endtime, 'H:mm', true, CENTRAL_TZ),
-    }))
-
-    let mealIndex = findIndex(times, ({startTime, endTime}) => now.isBetween(startTime, endTime))
-    if (mealIndex === undefined) {
-      if (now.isSameOrBefore(times[0].startTime)) {
-        mealIndex = 0
-      } else {
-        mealIndex = times.length - 1
-      }
-    }
-
-    return dayparts[0][mealIndex]
+    stationMenus: StationMenuType[],
+    foodItems: MenuItemContainerType,
+    menuLabel: string,
+    menuCorIcons: MasterCorIconMapType,
   }
 
   trimStationName(stationName: string) {
     return stationName.replace(/<strong>@(.*)<\/strong>/, '$1')
   }
 
-  parseData = async () => {
-    let {
-      now,
-      cafeId,
-      cafeMenu,
-      cafeInfo,
-      menuStations,
-      menuItems,
-      menuLabel,
-      menuCorIcons,
-    } = this.props
-
-    let foodItems = cafeMenu ? cafeMenu.items : menuItems
-    let mealName = menuLabel
-    let menus = menuStations
-    let icons = menuCorIcons
-
-    // if it's a bonapp menu
-    if (cafeMenu && cafeInfo && cafeId) {
-      let days = cafeInfo.cafes[cafeId].days
-      let today = days.find(({date}) => date === now.format('YYYY-MM-DD'))
-      if (!today || today.status === 'closed') {
-        this.setState({message: today ? today.message : 'Closed today'})
-        return
-      }
-
-      let dayparts = cafeMenu.days[0].cafes[cafeId].dayparts
-      let mealInfo = this.findMenu(dayparts, now)
-      menus = mealInfo ? mealInfo.stations : []
-      mealName = mealInfo ? mealInfo.label : ''
-      icons = cafeMenu.cor_icons
-    }
+  buildFilters = () => {
+    let {foodItems, menuCorIcons} = this.props
 
     let filters = []
     filters.push({
@@ -152,7 +65,7 @@ export class BaseMenuView extends React.Component {
       value: [],
     })
 
-    let allDietaryRestrictions = values(icons).map(item => item.label)
+    let allDietaryRestrictions = values(menuCorIcons).map(item => item.label)
     filters.push({
       type: 'list',
       multiple: true,
@@ -164,7 +77,7 @@ export class BaseMenuView extends React.Component {
       value: [],
     })
 
-    this.setState({menus, mealName, foodItems, filters})
+    this.setState({filters})
   }
 
   openFilterView = () => {
@@ -186,21 +99,17 @@ export class BaseMenuView extends React.Component {
   }
 
   render() {
-    if (this.state.message) {
-      return <LoadingView text={this.state.message} />
-    }
-
     return (
       <View style={{flex: 1, marginBottom: Platform.OS === 'ios' ? 49 : 0}}>
         <FilterToolbar
           date={this.props.now}
-          title={this.state.mealName}
+          title={this.props.menuLabel}
           appliedFilterCount={this.state.filters.filter(f => f.type === 'toggle' ? f.value : f.value.length > 0).length}
           onPress={this.openFilterView}
         />
         <FancyMenu
-          stationMenus={this.state.menus}
-          foodItems={this.state.foodItems}
+          stationMenus={this.props.stationMenus}
+          foodItems={this.props.foodItems}
           filters={this.state.filters}
           stationsToCreate={[]}
         />
